+++ hedera-mirror-web3/src/main/java/com/hedera/mirror/web3/state/core/ListReadableQueueState.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.hedera.mirror.web3.state.utils;
+package com.hedera.mirror.web3.state.core;
 
 import com.swirlds.state.spi.ReadableQueueStateBase;
 import jakarta.annotation.Nonnull;
