+++ hedera-mirror-web3/src/test/java/com/hedera/mirror/web3/state/core/MapReadableKVStateTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.hedera.mirror.web3.state.utils;
+package com.hedera.mirror.web3.state.core;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
